+++ src/devices/innr.ts
@@ -51,6 +51,13 @@ export const definitions: DefinitionWithExtend[] = [
         extend: [m.light({turnsOffAtBrightness1: true})],
     },
     {
+        zigbeeModel: ["AE 282 C"],
+        model: "AE 282 C",
+        vendor: "Innr",
+        description: "Smart Bulb Colour E27 (800lm)",
+        extend: [m.light({colorTemp: {range: [153, 556]}, color: {modes: ["xy", "hs"], enhancedHue: true}})],
+    },
+    {
         zigbeeModel: ["RCL 231 T"],
         model: "RCL 231 T",
         vendor: "Innr",
@@ -208,6 +215,13 @@ export const definitions: DefinitionWithExtend[] = [
         ota: true,
     },
     {
+        zigbeeModel: ["RB 252 C"],
+        model: "RB 252 C",
+        vendor: "Innr",
+        description: "Smart bulb Candle Colour E14",
+        extend: [m.light({colorTemp: {range: [153, 556]}, color: {modes: ["xy", "hs"], enhancedHue: true}})],
+    },
+    {
         zigbeeModel: ["RB 262"],
         model: "RB 262",
         vendor: "Innr",
@@ -324,6 +338,19 @@ export const definitions: DefinitionWithExtend[] = [
         description: "E27 bulb RGBW",
         extend: [m.light({colorTemp: {range: [153, 555]}, color: {modes: ["xy", "hs"], applyRedFix: true}, turnsOffAtBrightness1: true})],
         ota: true,
+    },
+    {
+        zigbeeModel: ["RB 287 C"],
+        model: "RB 287 C",
+        vendor: "Innr",
+        description: "Smart Bulb Colour E27 1200lm",
+        extend: [
+            m.light({
+                colorTemp: {range: [153, 556]},
+                color: {modes: ["xy", "hs"], enhancedHue: true},
+                turnsOffAtBrightness1: true,
+            }),
+        ],
     },
     {
         zigbeeModel: ["BY 285 C"],